refactor(grib1, grib2): both readers read through ByteSource like NetCDF does

+++ crates/fieldglass/src/session.rs
@@ -1180,7 +1180,7 @@ fn grib1_message(reader: &fieldglass_grib1::Grib1Reader, index: usize) -> Messag
         // Round-trips the `u32` handle `Session::message` was given and
         // `check_index` widened, so it cannot be a narrowing in practice.
         index: index as u32,
-        offset_bytes: msg.byte_offset as u64,
+        offset_bytes: msg.byte_offset,
         parameter,
         abbreviation,
         units,
@@ -1249,7 +1249,7 @@ fn grib2_message(reader: &fieldglass_grib2::Grib2Reader, index: usize) -> Messag
         // Round-trips the `u32` handle `Session::message` was given and
         // `check_index` widened, so it cannot be a narrowing in practice.
         index: index as u32,
-        offset_bytes: msg.byte_offset as u64,
+        offset_bytes: msg.byte_offset,
         parameter,
         abbreviation,
         units,
